More primitive type array2 implementations.
